Expose dataset conversion as the 'zea convert' subcommand
The converters were only reachable via 'python -m zea.data.convert', which is
inconsistent with the rest of the CLI (zea process, zea data ...) and not
discoverable from 'zea --help'. Add a top-level 'zea convert <dataset> <src>
<dst>' subcommand.

To keep 'zea --help' free of heavy imports, the convert CLI dataclasses move
from zea.data.convert.__main__ into zea.cli_args (importing zea.data eagerly
pulls in keras; zea.cli_args stays backend-free, matching how ProcessArgs and
DataArgs already work, with the heavy converter imported lazily in run()).
zea.data.convert.__main__ re-exports them so 'python -m zea.data.convert' and
existing imports keep working unchanged.

Conversion runs are dispatched with allow_preallocate=False, mirroring the
standalone entry point. Docs and tests updated for the new subcommand.
