+++ docs/source/cli.rst
@@ -2,7 +2,7 @@ Command line interface
 ================================
 
 Besides the main :doc:`zea API documentation <_autosummary/zea>`, ``zea`` provides a
-command line interface (CLI) with three primary subcommands (process, app, data).
+command line interface (CLI) with four primary subcommands (process, app, data, convert).
 
 Note that is very new functionality, and might change in future releases. Please report any issues you encounter.
 
@@ -11,13 +11,10 @@ Note that is very new functionality, and might change in future releases. Please
     zea process --dataset <path> --config <config.yaml> [options]  # batch beamform a dataset
     zea app [--share] [--server-port PORT]                         # launch the Gradio visualizer
     zea data <operation> [options]                                 # manipulate zea data files
+    zea convert <dataset> <src> <dst> [options]                    # convert raw datasets to zea
+
+The ``convert`` subcommand (documented below) is also available as a standalone
+module with an identical interface: ``python -m zea.data.convert <dataset> <src> <dst>``.
 
 .. tyroprogram:: zea.__main__:CLI
    :prog: zea
-
--------------------------------
-Convert datasets
--------------------------------
-
-.. tyroprogram:: zea.data.convert.__main__:Dataset
-   :prog: python -m zea.data.convert
